backend/backend/backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
flask-marshmallow
mysqlclient
marshmallow-sqlalchemy